Tue, Apr 12, 2022, 3:20 AM +00:00

+++ readability/readability.c
@@ -1,6 +1,7 @@
 #include <cs50.h>
 #include <stdio.h>
 #include <string.h>
+#include <math.h>
 
 int count_letters (string);
 int count_words (string);
@@ -17,7 +18,7 @@ int main(void)
 
 
     //calculation
-    float grade_score = 0.0588 * (count_letters / count_words * 100) - 0.296 * (count_sentences / count_words * 100) - 15.8);
+    float grade_score = 0.0588 * ((float)count_letters / (float)count_words * 100) - 0.296 * ((float)count_sentences / (float)count_words * 100) - 15.8);
     int output = round(grade_score);
